% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_testall.R
\name{debugonce_testall}
\alias{debugonce_testall}
\title{debugonce_testall
This function consumes an object of class \code{testall_summary} with the
ultimate goal to debug your function with a suggested specifications}
\usage{
debugonce_testall(sug, scenario = 1)
}
\arguments{
\item{sug}{an object of class \code{testall_summary}}

\item{scenario}{an optional numeric value indicating which scenario from sug$tests
will be used for debugging.}
}
\value{
Will take you directly into debug mode and will not return anything
}
\description{
debugonce_testall
This function consumes an object of class \code{testall_summary} with the
ultimate goal to debug your function with a suggested specifications
}
\examples{
# NOT RUN:
suggestion <- test_all(mean, list(x = c('asd')))
debugonce_testall(suggestion)
}
