% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_tester.R
\name{lazy_tester}
\alias{lazy_tester}
\title{lazy_tester
Create a grid corpus for \code{\link{test_all}}}
\usage{
lazy_tester(FUN, return_list = FALSE)
}
\arguments{
\item{FUN}{the function for which a grid corpus should be created}

\item{return_list}{a boolean which indicates whether a list should be returned}
}
\value{
Returns a grid output either in the console or as list
}
\description{
lazy_tester
Create a grid corpus for \code{\link{test_all}}
}
\examples{
# create a grid
lazy_tester(mean)
}
