% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.testall_summary.R, R/test_all.R
\name{print.testall_summary}
\alias{print.testall_summary}
\alias{test_all}
\title{test_all
A function which stress tests an R function with user-defined inputs.}
\usage{
\method{print}{testall_summary}(x, ...)

test_all(fun, input, output = NULL, cores = 1)
}
\arguments{
\item{x}{an object of class \code{testall_summary}}

\item{...}{arguments to be passed to the method}

\item{fun}{an R function to test}

\item{input}{a grid created with \code{\link{lazy_tester}}}

\item{output}{not implemented yet}

\item{cores}{an integer specifying the number of cores to use}
}
\value{
Returns a table with error information, i.e.
\itemize{
\item The function call (arguments)
\item In how many occurences it threw an error
}
}
\description{
test_all
A function which stress tests an R function with user-defined inputs.
}
\examples{
# NOT RUN:
my_sug <- test_all(mean, input = list(x = c('1','2','3')))
my_sug
# NOT RUN:
test_all(mean, list(x = c(1,2,3)))
}
